Added a sanitize names option to deal with source files that are named weirdly.
Sometimes the names of your source files are not really valid CSS selectors so if you treat then as such you're gonna have a bad time. Adding the :sanitizenames option makes it so the names are pre-processed to conform to CSS naming standards. NOTE: This is fundamentally incompatible with the colon and space escaping methods that were previously applied to the incoming file names.