Changed the assert_reference_image test harness to check for delta instead of equality.
It seems that some newer versions of imagemagick (or just it running on different OSs) changes the converted images very subtly. This change in the test harness makes it so that the images no longer have to be exactly equal, but need to be within a very close margin (currently 1e-7 or less percent difference) to pass the "equality" test.

Just for reference, in my case on a machine running OSX 10.9.2, MRI 2.1.1 & ImageMagick 6.8.7-7 Q16 the test_generate_sprite_from_other_formats test was within 7.8e-08 of exact. This was the only test that failed.